+++ default/role.harvester.js
@@ -2,6 +2,8 @@ module.exports = {
 
   run: creep => {
 
+    if (creep.suicideAt(20)) return;
+
     // At spawning Harvesters have 25% chances to be assigned to Towers too
     creep.memory.includeTowers = creep.memory.includeTowers ? creep.memory.includeTowers : _.sample([true, false, false]);