+++ default/role.hauler.js
@@ -2,6 +2,8 @@ module.exports = {
 
   run: creep => {
 
+    if (creep.suicideAt(20)) return;
+
     // creep.say("H");
     // if (creep.memory.includeTowers) creep.say("HT");
     // creep.memory.includeTowers = undefined; // RESET
@@ -24,8 +26,6 @@ module.exports = {
       If empty, look for the nearest fully charged storage or container
       */
     } else {
-
-      // creep.lookForAndPickupResource();
 
       const storage = creep.pos.findClosestByPath(FIND_STRUCTURES, {
         filter: s => ((